optimize getClinicalDataCounts query by using temp session scoped tables, and move related mapper tests to repo level
